008_best-ecommerce-review/requirements.txt
# Data Processing
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0

# Database
psycopg2-binary>=2.9.0  # PostgreSQL

# API & Web
requests>=2.31.0
python-dotenv>=1.0.0

# NLP & Scoring
konlpy>=0.6.0  # 한국어 형태소 분석
scikit-learn>=1.3.0

# 공개 데이터 수집
kaggle>=1.5.16                    # Kaggle 데이터셋
google-play-scraper>=1.2.4        # Google Play Store
tqdm>=4.65.0                      # 진행률 표시

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Logging
loguru>=0.7.0

# Security
cryptography>=41.0.0

# Utils
pyyaml>=6.0.0
python-dateutil>=2.8.0
